requirements.txt
Flask>=2.3.0
Flask-WTF>=1.2.0
Flask-Limiter>=3.5.0
Flask-SQLAlchemy>=3.1.0
python-dotenv>=1.0.0
requests>=2.31.0
Werkzeug>=2.3.0
WTForms>=3.1.0
